refactor: Transaction 처리 로직 추상화
