+++ app/src/main/java/com/techcourse/service/transaction/TransactionCallback.java
@@ -1,0 +1,7 @@
+package com.techcourse.service.transaction;
+
+@FunctionalInterface
+public interface TransactionCallback<T> {
+
+    T call();
+}
