+++ jdbc/src/main/java/nextstep/jdbc/JdbcCallback.java
@@ -3,6 +3,7 @@ package nextstep.jdbc;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
+@FunctionalInterface
 public interface JdbcCallback<T> {
 
     T call(PreparedStatement preparedStatement) throws SQLException;
